fix a couple of signed/unsigned comparison warnings
demonstrate using AssembleElemSolverAlgorithm in nso kernel test

Added a derivation of LinearSystem which plugs in and catches
the lhs and rhs that come through apply_coeff, so they can then
be compared to gold values as before.

Added new continuity-advection test which uses new ME methods...

This test uses the new AssembleElemSolverAlgorithmNewME class, which
does a slightly different scheme for copy/interleave of data with
respect to the new master-element API.

Made changes in ScratchViews, specifically a new
fill_master_element_views_new_ME method which makes the master-element
calls using views instead of pointers.

Currently disabled are SCS_GIJ and SCV_VOLUME, pending further
master-element work.

Add a pair of performance tests for cont-adv-elem-kernel.

One runs AssembleElemSolverAlgorithm, the other runs
AssembleElemSolverAlgorithmNewME which uses the new master-elem API.

vtune output shows that using the new master-elem API reduces the
cost of the master-element calls, as well as the copy/interleave
cost.

+++ unit_tests/UnitTestMasterElements.C
@@ -183,7 +183,7 @@ void check_exposed_face_shifted_ips_are_nodal(
 
   int index = 0;
   std::vector<std::vector<double>> shiftedIpList(shiftedIps.size() / dim);
-  for (unsigned j = 0; j < shiftedIps.size()/meSS.nDim_; ++j) {
+  for (int j = 0; j < (int)shiftedIps.size()/meSS.nDim_; ++j) {
     shiftedIpList.at(j).resize(dim);
     for (int d = 0; d < dim; ++d) {
       shiftedIpList.at(j).at(d) = shiftedIps[index];
